sort atom id in dumps of SNA descriptors
By default, dump files of LAMMPs has an indeterminate order for atoms. To constrain the sequence of atom to be consistent with that of a structure, "dump_modify 1 sort id" should be added for LAMMPS input files. An alternative way is to also dump the atom ids into the dump files, so that the calculated atomic descriptors can be correctly allocated to atoms. 

In calculator templates of maml,  "dump_modify 1 sort id" is used consistently, and only the "SpectralNeighborAnalysis" misses this line of command. So in this commit, this dump_modify is added to "SpectralNeighborAnalysis".

Interestingly, for small cells, LAMMPS dumps the atoms with sorted id by default, while the ids are unsorted for large cells. This behavior is seemingly not described in LAMMPS documentation, and it is unclear where is the border for "large" and "small". This is why that our previous works on SNAP did not have problem, as the dumped atoms for the small training structures were likely sorted by default.